Disable casting float64 init to float32 (#64)
* [Core] move away from default fix_float64 behavior

now set to False by default, and carried as part of ModelWrapper

* [Refactor] specify dtypes for a more np array creation code

* [Lint] run pre-commit on all files

* [Docs] remove bullets to resolve docs build error

+++ src/finn/transformation/remove.py
@@ -101,7 +101,7 @@ class RemoveIdentityOps(Transformation):
                 and not model.is_join_node(n)
             ):
                 pads = get_by_name(n.attribute, "pads")
-                pads = np.asarray(pads.ints)
+                pads = np.asarray(pads.ints, dtype=np.int64)
                 if (pads == 0).all():
                     remove_node_and_rewire(model, n)
                     graph_modified = True
